fix: domain が data に依存するのを防ぐため、変換メソッドを data に置いた
