+++ app/src/main/java/app/doggy/newmybrary/data/repository/BookRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package app.doggy.newmybrary.data.repository
 
+import app.doggy.newmybrary.data.api.response.toBook
 import app.doggy.newmybrary.data.api.service.BookApi
 import app.doggy.newmybrary.domain.model.Book
-import app.doggy.newmybrary.domain.model.toBook
 import app.doggy.newmybrary.domain.repository.BookRepository
 import dagger.Binds
 import dagger.Module
