+++ app/src/main/java/app/doggy/newmybrary/domain/model/Book.kt
@@ -1,6 +1,5 @@
 package app.doggy.newmybrary.domain.model
 
-import app.doggy.newmybrary.data.api.response.FetchBooksResponseItem
 import java.util.Date
 
 data class Book(
@@ -12,15 +11,4 @@ data class Book(
   val imageUrl: String,
   val recordList: List<Diary>,
   val registeredAt: Date?,
-)
-
-fun FetchBooksResponseItem.toBook() = Book(
-  id = this.id,
-  title = this.volumeInfo.title,
-  author = this.volumeInfo.authors,
-  description = this.volumeInfo.description,
-  totalPage = this.volumeInfo.pageCount,
-  imageUrl = this.volumeInfo.imageLinks.thumbnail,
-  recordList = listOf(),
-  registeredAt = null
 )
